unified scripts & added justfile
