+++ .github/actions/detect-changes/action.yml
@@ -1,0 +1,33 @@
+name: detect-changes
+description: Detect whether files under a given path changed in this PR/push. Sets the output `has_changed` to "true" or "false".
+inputs:
+  path:
+    description: "Path prefix to check (e.g. 'ray', 'backend'). Matched as '^<path>/'."
+    required: true
+outputs:
+  has_changed:
+    description: '"true" if files under the path changed, else "false".'
+    value: ${{ steps.detect.outputs.has_changed }}
+runs:
+  using: composite
+  steps:
+    - name: Detect changes
+      id: detect
+      shell: bash
+      env:
+        TARGET_PATH: ${{ inputs.path }}
+      run: |
+        set -euo pipefail
+        has_changed=false
+        if [ "${{ github.event_name }}" = "pull_request" ]; then
+          git fetch origin ${{ github.event.pull_request.base.ref }}
+          base_branch="origin/${{ github.event.pull_request.base.ref }}"
+          if git diff --name-only "$base_branch" HEAD | grep -q -e "^${TARGET_PATH}/"; then
+            has_changed=true
+          fi
+        else
+          if git diff --name-only ${{ github.event.before }} ${{ github.sha }} | grep -q -e "^${TARGET_PATH}/"; then
+            has_changed=true
+          fi
+        fi
+        echo "has_changed=${has_changed}" >> "$GITHUB_OUTPUT"
