+++ .github/workflows/ray_checks.yml
@@ -18,15 +18,12 @@ jobs:
       - name: Checkout
         uses: actions/checkout@v7
       - name: Prepare environment
-        run: |
-          ./bin/setup-folders.sh
-          ./bin/setup-envs.sh --project_name action-runner --port_prefix 131
-          sed -i 's/\(DATS_RAY_DOCKER_VERSION=[0-9.]*\)/\1-${{ github.run_id }}/' docker/.env
-          sed -i 's/^RAY_DEVICE_IDS=.*/RAY_DEVICE_IDS=1/' docker/.env
-          echo "RAY_CACHE_DIR=$HOME/ray_cache" >> docker/.env
-          mkdir -p $HOME/ray_cache
-          chmod 777 $HOME/ray_cache
-          sed -i 's/ray_network/ray_network_${{ github.run_id }}/g' docker/compose.ray.yml
+        uses: ./.github/actions/prepare-env
+        with:
+          ray: true
+          hf_token: ${{ secrets.HF_HUB_TOKEN }}
+      - name: Isolate ray network for this run
+        run: sed -i 's/ray_network/ray_network_${{ github.run_id }}/g' docker/compose.ray.yml
       - name: Build & Start Docker Containers
         working-directory: docker
         run: |
@@ -34,9 +31,7 @@ jobs:
           docker compose -f compose.ray.yml build
           docker compose -f compose.ray.yml up --wait
       - name: Cleanup
-        working-directory: docker
         if: always()
-        run: |
-          docker compose -f compose.ray.yml down -v
-          RAY_IMAGE=uhhlt/dats_ray:$(grep -oP 'DATS_RAY_DOCKER_VERSION=\K.*' .env)
-          docker rmi $RAY_IMAGE
+        uses: ./.github/actions/cleanup
+        with:
+          ray: true
