+++ .vscode/dats.code-workspace
@@ -81,6 +81,8 @@
       // frontend
       "dbaeumer.vscode-eslint",
       "esbenp.prettier-vscode",
+      // just
+      "nefrob.vscode-just-syntax"
     ],
   },
 }
