+++ README.md
@@ -66,11 +66,11 @@ git clone https://github.com/uhh-lt/dats.git
 #### 2. Run setup scripts
 
 ```bash
-./bin/setup-envs.sh --project_name dats --port_prefix 101
+./bin/setup/setup-envs.sh --project_name dats --port_prefix 101
 ```
 
 ```bash
-./bin/setup-folders.sh
+./bin/setup/setup-folders.sh
 ```
 
 #### 3. Start Docker containers
