+++ backend/.env.example
@@ -51,3 +51,7 @@ DOCLING_PORT=13139
 
 GLITCHTIP_DSN_BACKEND=
 GLITCHTIP_DSN_FRONTEND=
+
+# Debugpy listen ports for local debugging.
+DEBUGPY_PORT_API=13190
+DEBUGPY_PORT_WORKER=13191
