+++ backend/.vscode/launch.json
@@ -4,95 +4,36 @@
     {
       "name": "fastapi",
       "type": "debugpy",
-      "request": "launch",
-      "preLaunchTask": "Run DATS Setup",
-      "module": "uvicorn",
-      "args": ["main:app", "--reload", "--host", "0.0.0.0", "--app-dir", "src"],
-      "console": "integratedTerminal",
-      "justMyCode": true,
-      "cwd": "${workspaceFolder}",
-      "envFile": "${workspaceFolder}/.env",
-      "env": {
-        "PYTHONPATH": "${workspaceFolder}/src"
-      }
+      "request": "attach",
+      "preLaunchTask": "dev: backend-api (debug)",
+      "connect": {
+        "host": "localhost",
+        "port": "${env:DEBUGPY_PORT_API}"
+      },
+      "pathMappings": [
+        {
+          "localRoot": "${workspaceFolder}",
+          "remoteRoot": "."
+        }
+      ],
+      "justMyCode": true
     },
     {
       "name": "rq",
       "type": "debugpy",
-      "request": "launch",
-      "program": "${workspaceFolder}/src/worker.py",
-      "args": ["work", "dev"],
-      "console": "integratedTerminal",
-      "justMyCode": true,
-      "cwd": "${workspaceFolder}",
-      "envFile": "${workspaceFolder}/.env",
-      "env": {
-        "PYTHONPATH": "${workspaceFolder}/src",
-        "PATH": "${workspaceFolder}/.venv/bin:${env:PATH}",
-        "RQ_WORKERS_CPU": "1",
-        "RQ_WORKERS_API": "1",
-        "RQ_WORKERS_GPU": "1"
-      }
-    },
-    {
-      "name": "pytest",
-      "type": "debugpy",
-      "request": "launch",
-      "module": "pytest",
-      "console": "integratedTerminal",
-      "justMyCode": true,
-      "cwd": "${workspaceFolder}",
-      "envFile": "${workspaceFolder}/.env",
-      "env": {
-        "PYTHONPATH": "${workspaceFolder}/src",
-        "RESET_DATABASE_FOR_TESTING": "1"
-      }
-    },
-    {
-      "name": "pyright",
-      "type": "node-terminal",
-      "request": "launch",
-      "command": "uv run pyright",
-      "cwd": "${workspaceFolder}",
-      "envFile": "${workspaceFolder}/.env",
-      "env": {
-        "PYTHONPATH": "${workspaceFolder}/src"
-      }
-    },
-    {
-      "name": "Alembic: migrate",
-      "type": "debugpy",
-      "request": "launch",
-      "program": "${workspaceFolder}/src/migrations/run_migrations.py",
-      "console": "integratedTerminal",
-      "justMyCode": true,
-      "cwd": "${workspaceFolder}",
-      "envFile": "${workspaceFolder}/.env",
-      "env": {
-        "PYTHONPATH": "${workspaceFolder}/src"
-      }
-    },
-    {
-      "name": "Alembic: check",
-      "type": "node-terminal",
-      "request": "launch",
-      "command": "uv run alembic -c src/migrations/alembic.ini check",
-      "cwd": "${workspaceFolder}",
-      "envFile": "${workspaceFolder}/.env",
-      "env": {
-        "PYTHONPATH": "${workspaceFolder}/src"
-      }
-    },
-    {
-      "name": "Alembic: revision",
-      "type": "node-terminal",
-      "request": "launch",
-      "command": "uv run alembic -c src/migrations/alembic.ini revision --autogenerate -m \"vscode launcher\"",
-      "cwd": "${workspaceFolder}",
-      "envFile": "${workspaceFolder}/.env",
-      "env": {
-        "PYTHONPATH": "${workspaceFolder}/src"
-      }
+      "request": "attach",
+      "preLaunchTask": "dev: backend-worker (debug)",
+      "connect": {
+        "host": "localhost",
+        "port": "${env:DEBUGPY_PORT_WORKER}"
+      },
+      "pathMappings": [
+        {
+          "localRoot": "${workspaceFolder}",
+          "remoteRoot": "."
+        }
+      ],
+      "justMyCode": true
     }
   ]
 }
