+++ bin/prod/backup-weaviate.sh
