+++ bin/prod/pg_max_pool_sanity_check.sh
