+++ bin/prod/restore-elasticsearch.sh
