+++ bin/prod/restore-weaviate.sh
