+++ docker/.env.backend.example
@@ -56,7 +56,7 @@ ASSISTANT_USER_PASSWOD=12ASSISTANT34
 #   total                   = 120 + 50          = 170 <= 200  (30 headroom)
 #
 # Before changing any of these values, run:
-#   bin/pg_max_pool_sanity_check.sh
+#   bin/prod/pg_max_pool_sanity_check.sh
 # It reads the real values from .env / .env.backend / compose.yml and fails if
 # the invariant above would be violated.
 POSTGRES_POOL_SIZE=8
