+++ docker/.env.example
@@ -61,7 +61,7 @@ GLITCHTIP_DSN_FRONTEND=
 #   NUM_CONNECTIONS_WORKERS = (RQ_WORKERS_CPU + RQ_WORKERS_API + RQ_WORKERS_GPU) * 2
 #   NUM_CONNECTIONS_API + NUM_CONNECTIONS_WORKERS <= POSTGRES_MAX_CONNECTIONS
 #
-# Full breakdown in .env.backend. Verify with bin/pg_max_pool_sanity_check.sh.
+# Full breakdown in .env.backend. Verify with bin/prod/pg_max_pool_sanity_check.sh.
 API_WORKERS=10
 API_EXPOSED=13120
 # Logging ('critical', 'error', 'warning', 'info', 'debug', 'trace')
@@ -90,7 +90,7 @@ POSTGRES_PASSWORD=dats123
 #   NUM_CONNECTIONS_API + NUM_CONNECTIONS_WORKERS <= POSTGRES_MAX_CONNECTIONS
 #   NUM_CONNECTIONS_API     = API_WORKERS * (POSTGRES_POOL_SIZE + POSTGRES_MAX_OVERFLOW)
 #   NUM_CONNECTIONS_WORKERS = (RQ_WORKERS_CPU + RQ_WORKERS_API + RQ_WORKERS_GPU) * 2
-# Full breakdown in .env.backend. Verify with bin/pg_max_pool_sanity_check.sh.
+# Full breakdown in .env.backend. Verify with bin/prod/pg_max_pool_sanity_check.sh.
 POSTGRES_MAX_CONNECTIONS=200
 
 ######################
