+++ frontend/package.json
@@ -6,15 +6,9 @@
   "scripts": {
     "dev": "vite",
     "build": "tsc && vite build",
-    "lint": "eslint . --max-warnings 0",
-    "lint-fix": "eslint . --ext js,jsx,ts,tsx --max-warnings 0 --fix",
-    "format": "prettier --write --ignore-unknown --config ../.prettierrc.yaml --ignore-path ../.prettierignore .",
-    "tsc-watch": "tsc --noEmit --watch",
-    "typecheck": "tsc --noEmit",
     "preview": "vite preview",
     "update-api": "node bin/updateAPI.js",
     "update-emojis": "node bin/updateEmojiData.js",
-    "generate-api": "node bin/generateAPI.js",
     "update-version": "node bin/updateVersion.js"
   },
   "dependencies": {
