+++ ray/.env.example
@@ -8,3 +8,6 @@ CUDA_VISIBLE_DEVICES=0
 
 # Glitchtip DSN connection
 GLITCHTIP_DSN=
+
+# Debugpy listen port for local debugging (attach).
+DEBUGPY_PORT_RAY=13192
