+++ ray/.vscode/launch.json
@@ -2,40 +2,21 @@
   "version": "0.2.0",
   "configurations": [
     {
-      "name": "serve",
-      "type": "node-terminal",
-      "request": "launch",
-      "command": "${workspaceFolder}/src/ray_entrypoint_dev.sh",
-      "cwd": "${workspaceFolder}/src",
-      "envFile": "${workspaceFolder}/.env",
-      "env": {
-        "PYTHONPATH": "${workspaceFolder}/src"
-      }
-    },
-    {
-      "name": "pytest",
+      "name": "ray",
       "type": "debugpy",
-      "request": "launch",
-      "module": "pytest",
-      "console": "integratedTerminal",
-      "justMyCode": true,
-      "cwd": "${workspaceFolder}",
-      "envFile": "${workspaceFolder}/.env",
-      "env": {
-        "PYTHONPATH": "${workspaceFolder}/src",
-        "RESET_DATABASE_FOR_TESTING": "1"
-      }
-    },
-    {
-      "name": "pyright",
-      "type": "node-terminal",
-      "request": "launch",
-      "command": "uv run pyright",
-      "cwd": "${workspaceFolder}",
-      "envFile": "${workspaceFolder}/.env",
-      "env": {
-        "PYTHONPATH": "${workspaceFolder}/src"
-      }
+      "request": "attach",
+      "preLaunchTask": "dev: ray (debug)",
+      "connect": {
+        "host": "localhost",
+        "port": "${env:DEBUGPY_PORT_RAY}"
+      },
+      "pathMappings": [
+        {
+          "localRoot": "${workspaceFolder}",
+          "remoteRoot": "."
+        }
+      ],
+      "justMyCode": true
     }
   ]
 }
