+++ ray/update_version.py
@@ -1,0 +1,36 @@
+import argparse
+import os
+
+parser = argparse.ArgumentParser(description="DATS ray release script")
+parser.add_argument(
+    "--version",
+    type=str,
+    help="New version to release",
+    required=True,
+    dest="version",
+)
+args = parser.parse_args()
+
+root_dir = os.path.dirname(__file__)
+
+# Update pyproject.toml
+with open(os.path.join(root_dir, "pyproject.toml"), "r") as f:
+    pyproject = f.readlines()
+for i, line in enumerate(pyproject):
+    if line.startswith("version = "):
+        pyproject[i] = f'version = "{args.version}"\n'
+        break
+with open(os.path.join(root_dir, "pyproject.toml"), "w") as f:
+    f.writelines(pyproject)
+
+# Update uv.lock
+with open(os.path.join(root_dir, "uv.lock"), "r") as f:
+    uv_lock = f.readlines()
+previous_line = ""
+for i, line in enumerate(uv_lock):
+    if previous_line.startswith('name = "dats-ray"') and line.startswith("version = "):
+        uv_lock[i] = f'version = "{args.version}"\n'
+        break
+    previous_line = line
+with open(os.path.join(root_dir, "uv.lock"), "w") as f:
+    f.writelines(uv_lock)
